Fix mfcc computation on new wavs

+++ egs/wsj/s5/utils/compute_mfcc_new_wavs.sh
@@ -16,9 +16,26 @@ dir=$1
 
 mkdir -p $dir
 
-cat $dir/wav.scp | cut -d' ' -f1 > $dir/utt_ids
-paste $dir/utt_ids $dir/utt_ids  > $dir/spk2utt
-cp $dir/spk2utt $dir/utt2spk
+if [ ! -f $dir/spk2utt ]; then
+    if [ -f $dir/utt2spk ]; then
+	utils/utt2spk_to_spk2utt.pl $dir/utt2spk > $dir/spk2utt
+    fi
+fi
+
+if [ ! -f $dir/utt2spk ]; then
+    if [ -f $dir/spk2utt ]; then
+	utils/spk2utt_to_utt2spk.pl $dir/spk2utt > $dir/utt2spk
+    fi
+fi
+
+if [ ! -f $dir/spk2utt ]; then
+    >&2 echo "Generating spk2utt and utt2spk"
+    cat $dir/wav.scp | cut -d' ' -f1 > $dir/utt_ids
+    paste -d' ' $dir/utt_ids $dir/utt_ids  > $dir/spk2utt
+    cp $dir/spk2utt $dir/utt2spk
+else
+    >&2 echo "Not generating spk2utt and utt2spk since they are already there"
+fi
 
 cat conf/mfcc.conf $dir/conf/mfcc.conf 2>/dev/null > $dir/mfcc.conf
